refactor(study-session): colocate composables + lift out of modals/
Move the study-session composables (session-core, flashcard-session,
card-edit, card-actions, card-preview, session-cards, study-modal) into
components/study-session/composables/, colocated with the feature view —
same convention as the deck card-editor. Also lift the feature out of
components/modals/ to components/study-session/, since it's a feature
folder, not a generic modal primitive.
