+++ src/components/study-session/composables/session-core.ts
