+++ src/components/study-session/session-flashcard/finish-animation.vue
