+++ src/components/study-session/session-flashcard/index.vue
@@ -4,11 +4,11 @@ import CardStage from './card-stage.vue'
 import StudyEditFooter from './study-edit-footer.vue'
 import RatingButtons from './rating-buttons.vue'
 import FinishAnimation from './finish-animation.vue'
-import { useFlashcardSession } from '@/composables/study-session/flashcard-session'
-import { useCardPreview } from '@/composables/study-session/card-preview'
-import { useCardEdit } from '@/composables/study-session/card-edit'
-import { useActiveCardActions } from '@/composables/study-session/card-actions'
-import { useSessionCards } from '@/composables/study-session/session-cards'
+import { useFlashcardSession } from '@/components/study-session/composables/flashcard-session'
+import { useCardPreview } from '@/components/study-session/composables/card-preview'
+import { useCardEdit } from '@/components/study-session/composables/card-edit'
+import { useActiveCardActions } from '@/components/study-session/composables/card-actions'
+import { useSessionCards } from '@/components/study-session/composables/session-cards'
 import { useModalRequestClose } from '@/composables/modal'
 import { type Grade } from 'ts-fsrs'
 import { computed, useTemplateRef } from 'vue'
